works with overlapping calls to hydec
Former-commit-id: 50e834c07f7ea3117093f96298c723b0cb9dad2c [formerly c2b22ae7a906a9eb36362675243867828718b5f1]
Former-commit-id: 73070a6093291de15618f4abdca023156ace9b7b

+++ src/mitll/langtest/client/scoring/AudioPanel.java
@@ -375,7 +375,7 @@ public class AudioPanel extends VerticalPanel implements RequiresResize {
       int offsetHeight = imageContainer.getOffsetHeight();
       imageOverlay.setSize("2px", offsetHeight + "px");
 
-      int pixelProgress = (int) (((double) imageContainer.getOffsetWidth()) * (position / songDurationInMillis));
+      int pixelProgress = (int) (((float) imageContainer.getOffsetWidth()) * ((float)position / (float)songDurationInMillis));
       int left = imageContainer.getAbsoluteLeft() + pixelProgress;
       int top = imageContainer.getAbsoluteTop();
       imageOverlay.setPopupPosition(left, top);
